+++ src/error.rs
@@ -7,6 +7,8 @@ use std::fmt;
 use svgparser::Error as ParseError;
 use svgparser::ErrorPos;
 
+use simplecss::Error as CssParseError;
+
 /// List of all errors that can occur during processing of the SVG DOM.
 #[derive(PartialEq)]
 pub enum Error {
@@ -27,6 +29,8 @@ pub enum Error {
     ElementCrosslink,
     /// Error from *libsvgparser*.
     ParseError(ParseError),
+    /// Error from *simplecss*.
+    CssParseError(CssParseError),
     /// Parsed document must have an `svg` element.
     NoSvgElement,
     /// Parsed document must have at least one node.
@@ -45,21 +49,31 @@ pub enum Error {
     /// <rect fill="url(#lg1) red"/>
     /// ```
     UnsupportedPaintFallback(String), // FuncIRI name
-    /// We don't support `use` elements with broken filter attribute.
+    /// We don't support `use` elements with a broken filter attribute.
     BrokenFuncIri(String), // FuncIRI name
 }
 
 impl fmt::Display for Error {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
         match *self {
-            Error::ElementMustHaveAnId => write!(f, "Element must have an id"),
-            Error::ElementCrosslink => write!(f, "Element crosslink"),
-            Error::ParseError(e) => write!(f, "{:?}", e),
-            Error::NoSvgElement => write!(f, "Document didn't have an SVG element"),
-            Error::EmptyDocument => write!(f, "Document didn't have any nodes"),
-            Error::UnsupportedCSS(ref pos) => write!(f, "Unsupported CSS at {:?}", pos),
-            Error::InvalidCSS(ref pos) => write!(f, "Invalid CSS at {:?}", pos),
-            Error::UnsupportedEntity(ref pos) => write!(f, "Unsupported ENTITY data at {:?}", pos),
+            Error::ElementMustHaveAnId =>
+                write!(f, "Element must have an id"),
+            Error::ElementCrosslink =>
+                write!(f, "Element crosslink"),
+            Error::ParseError(e) =>
+                write!(f, "{:?}", e),
+            Error::CssParseError(e) =>
+                write!(f, "{:?}", e),
+            Error::NoSvgElement =>
+                write!(f, "Document didn't have an SVG element"),
+            Error::EmptyDocument =>
+                write!(f, "Document didn't have any nodes"),
+            Error::UnsupportedCSS(ref pos) =>
+                write!(f, "Unsupported CSS at {:?}", pos),
+            Error::InvalidCSS(ref pos) =>
+                write!(f, "Invalid CSS at {:?}", pos),
+            Error::UnsupportedEntity(ref pos) =>
+                write!(f, "Unsupported ENTITY data at {:?}", pos),
             Error::UnsupportedPaintFallback(ref iri) =>
                 write!(f, "Valid FuncIRI(#{}) with fallback value is not supported", iri),
             Error::BrokenFuncIri(ref iri) =>
@@ -78,5 +92,11 @@ impl fmt::Debug for Error {
 impl From<ParseError> for Error {
     fn from(value: ParseError) -> Error {
         Error::ParseError(value)
+    }
+}
+
+impl From<CssParseError> for Error {
+    fn from(value: CssParseError) -> Error {
+        Error::CssParseError(value)
     }
 }